pyproject.toml
[tool.poetry]
name = "supply-chain-ai"
version = "1.0.0"
description = "Multi-Agent AI-Powered Supply Chain Orchestration Platform"
authors = ["Supply Chain AI Team <team@supply-chain.ai>"]
readme = "README.md"
python = "^3.11"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
anthropic = "^0.40.0"
pydantic = "^2.6.0"
pydantic-settings = "^2.2.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.27"}
asyncpg = "^0.29.0"
alembic = "^1.13.1"
redis = {extras = ["hiredis"], version = "^5.0.1"}
aiokafka = "^0.10.0"
httpx = "^0.26.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
structlog = "^24.1.0"
prometheus-fastapi-instrumentator = "^6.1.0"
opentelemetry-api = "^1.22.0"
opentelemetry-sdk = "^1.22.0"
opentelemetry-instrumentation-fastapi = "^0.43b0"
opentelemetry-exporter-jaeger = "^1.21.0"
mlflow = "^2.10.0"
scikit-learn = "^1.4.0"
xgboost = "^2.0.3"
pandas = "^2.2.0"
numpy = "^1.26.0"
boto3 = "^1.34.0"
tenacity = "^8.2.3"
celery = {extras = ["redis"], version = "^5.3.6"}
python-dotenv = "^1.0.0"
aiofiles = "^23.2.1"
pytz = "^2024.1"
email-validator = "^2.1.0"
slowapi = "^0.1.9"
prometheus-client = "^0.20.0"
shap = "^0.44.1"
optuna = "^3.5.0"

[tool.poetry.dev-dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.4"
pytest-cov = "^4.1.0"
httpx = "^0.26.0"
fakeredis = "^2.21.1"
black = "^24.1.1"
ruff = "^0.2.1"
mypy = "^1.8.0"
locust = "^2.24.0"
factory-boy = "^3.3.0"
Faker = "^23.2.1"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "B008", "B904"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "smoke: Smoke tests",
    "slow: Slow tests",
]
addopts = "-v --tb=short --strict-markers"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
